Move components/SubforestColumns -> components/SubforestTables/Columns
